+++ webpack.config.js
@@ -1,16 +1,22 @@
 const path = require('path');
 const webpack = require('webpack');
 const WebpackAutoInject = require('webpack-auto-inject-version');
+const nodeExternals = require('webpack-node-externals');
 
 module.exports = (env) => {
   const minimize = env && env.production;
   const cdn = env && env.cdn;
 
   let filename = 'streaming-client';
+  let mode;
+  let externals = [];
   let babelExcludes = [];
+  let babelOptions;
 
   /* if building for the cdn */
   if (cdn) {
+    filename += '.bundle';
+
     /*
       this is so babel doesn't try to polyfill/transpile core-js (which is the polyfill)
         and the build tools.
@@ -21,21 +27,59 @@ module.exports = (env) => {
       /\bwebpack\/buildin\b/
     ];
 
-    filename += '.bundle';
+    babelOptions = {
+      sourceType: 'unambiguous',
+      ignore: [/\/core-js/],
+      presets: [
+        ['@babel/preset-env',
+          {
+            debug: false, /* set to `true` if you start banging your head against the wall */
+            targets: [
+              'last 2 versions',
+              '> 5%',
+              'IE 11',
+              'not dead'
+            ]
+          }
+        ]
+      ],
+      plugins: [
+        ['@babel/plugin-transform-runtime', {
+          corejs: 3
+        }]
+      ]
+    };
   } else {
-    /* if we are building for 'module', don't polyfill/transpile any dependencies */
-    babelExcludes = [/node_modules/];
+    /* we are building for node */
+    const modulesToBundle = [/* left this here just in case it is needed */];
+
+    /* we don't want to bundle node_modules */
+    externals.push(nodeExternals({
+      allowlist: modulesToBundle.map(m => new RegExp(`^${m.replace(/-/g, '\\-')}`))
+    }));
+
+    /* if we are building for 'module', don't polyfill/transpile most dependencies */
+    babelExcludes = [
+      new RegExp(`/node_modules/(?!${modulesToBundle.join('|')}/)`)
+    ];
+
+    babelOptions = {
+      sourceType: 'unambiguous',
+      presets: ['@babel/preset-env']
+    };
   }
 
   filename += minimize ? '.min.js' : '.js';
+  mode = minimize ? 'production' : 'development';
 
   return {
-    target: cdn ? 'web' : 'node',
+    target: 'web',
     entry: './src/client.js',
-    mode: minimize ? 'production' : 'development',
+    mode,
     optimization: {
       minimize
     },
+    externals,
     devtool: minimize ? 'source-map' : '',
     output: {
       path: path.resolve(__dirname, 'dist'),
@@ -61,7 +105,8 @@ module.exports = (env) => {
         {
           test: /\.(c|m)?js$/,
           loader: 'babel-loader',
-          exclude: babelExcludes
+          exclude: babelExcludes,
+          options: babelOptions
         }
       ]
     }