+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased](https://github.com/purecloudlabs/purecloud-streaming-client/compare/v11.3.1...HEAD)
+### Changed
+* updated auth failure log to pass channelId into the details object - [PR#57](https://github.com/purecloudlabs/purecloud-streaming-client/pull/57)
 
 # [v11.3.1](https://github.com/purecloudlabs/purecloud-streaming-client/compare/v11.3.0...v11.3.1)
 ### Changed